feat: add logic to parse gl xml
