--- ci/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "ci"
-version = "0.1.0"
-authors = ["Martin Krošlák <kroslakma@gmail.com>"]
-
-[dependencies]
-
-[dev-dependencies]
-minreq = { version = "^2", features = ["https"] }
-ron = "^0.4"
-serde = "^1.0.75"
-serde_derive = "^1.0.75"
-vk-parse = { path = "../vk-parse", features = ["serialize", "vkxml-convert"] }
-vkxml = "^0.3"
-xml-rs = "^0.8"
