+++ vk-parse/src/util.rs
@@ -1,0 +1,279 @@
+use std::io::Read;
+use std::str::FromStr;
+use crate::types::Error;
+use xml::reader::XmlEvent;
+
+pub type XmlEvents<R> = xml::reader::Events<R>;
+pub type XmlAttribute = xml::attribute::OwnedAttribute;
+
+pub fn xpath_attribute(xpath: &str, attribute_name: &str) -> String {
+    let mut xpath = String::from(xpath);
+    xpath.push_str("[@");
+    xpath.push_str(attribute_name);
+    xpath.push(']');
+    xpath
+}
+
+
+//--------------------------------------------------------------------------------------------------
+#[macro_export]
+macro_rules! unwrap_attribute (
+    ($ctx:expr, $element:ident, $attribute:ident) => {
+        let $attribute = match $attribute {
+            Some(val) => val,
+            None => {
+                $ctx.errors.push(Error::MissingAttribute {
+                    xpath: $ctx.xpath.clone(),
+                    name: String::from(stringify!($attribute)),
+                });
+                return None;
+            }
+        };
+    };
+);
+
+#[macro_export]
+macro_rules! match_attributes {
+    ($ctx:expr, $a:ident in $attributes:expr, $($p:pat => $e:expr),+) => {
+        for $a in $attributes {
+            let n = $a.name.local_name.as_str();
+            match n {
+                $(
+                    $p => $e,
+                )+
+                _ => $ctx.errors.push(Error::UnexpectedAttribute {
+                    xpath: $ctx.xpath.clone(),
+                    name: String::from(n),
+                })
+            }
+        }
+    };
+}
+
+#[macro_export]
+macro_rules! match_elements {
+    ($ctx:expr, $($p:pat => $e:expr),+) => {
+        while let Some(Ok(e)) = $ctx.events.next() {
+            match e {
+                XmlEvent::StartElement { name, .. } => {
+                    let name = name.local_name.as_str();
+                    $ctx.push_element(name);
+                    match name {
+                        $(
+                            $p => $e,
+                        )+
+                        _ => {
+                            $ctx.errors.push(Error::UnexpectedElement {
+                                xpath: $ctx.xpath.clone(),
+                                name: String::from(name),
+                            });
+                            consume_current_element($ctx);
+                        }
+                    }
+                }
+                XmlEvent::EndElement { .. } => {
+                    $ctx.pop_element();
+                    break;
+                }
+                _ => {}
+            }
+        }
+    };
+
+    ( $ctx:expr, $attributes:ident, $($p:pat => $e:expr),+) => {
+        while let Some(Ok(e)) = $ctx.events.next() {
+            match e {
+                XmlEvent::StartElement { name, $attributes, .. } => {
+                    let name = name.local_name.as_str();
+                    $ctx.push_element(name);
+                    match name {
+                        $(
+                            $p => $e,
+                        )+
+                        _ => {
+                            $ctx.errors.push(Error::UnexpectedElement {
+                                xpath: $ctx.xpath.clone(),
+                                name: String::from(name),
+                            });
+                            consume_current_element($ctx);
+                        }
+                    }
+                }
+                XmlEvent::EndElement { .. } => {
+                    $ctx.pop_element();
+                    break;
+                }
+                _ => {}
+            }
+        }
+    };
+}
+
+#[macro_export]
+macro_rules! match_elements_combine_text {
+    ( $ctx:expr, $buffer:ident, $($p:pat => $e:expr),+) => {
+        while let Some(Ok(e)) = $ctx.events.next() {
+            match e {
+                XmlEvent::Characters(text) => $buffer.push_str(&text),
+                XmlEvent::Whitespace(text) => $buffer.push_str(&text),
+                XmlEvent::StartElement { name, .. } => {
+                    $buffer.push(' ');
+                    let name = name.local_name.as_str();
+                    $ctx.push_element(name);
+                    match name {
+                        $(
+                            $p => $e,
+                        )+
+                        _ => {
+                            $ctx.errors.push(Error::UnexpectedElement {
+                                xpath: $ctx.xpath.clone(),
+                                name: String::from(name),
+                            });
+                            consume_current_element($ctx);
+                        }
+                    }
+                }
+                XmlEvent::EndElement { .. } => {
+                    $buffer.push(' ');
+                    $ctx.pop_element();
+                    break;
+                },
+                _ => {}
+            }
+        }
+    };
+
+    ( $ctx:expr, $attributes:ident, $buffer:ident, $($p:pat => $e:expr),+) => {
+        while let Some(Ok(e)) = $ctx.events.next() {
+            match e {
+                XmlEvent::Characters(text) => $buffer.push_str(&text),
+                XmlEvent::Whitespace(text) => $buffer.push_str(&text),
+                XmlEvent::StartElement { name, $attributes, .. } => {
+                    let name = name.local_name.as_str();
+                    $ctx.push_element(name);
+                    match name {
+                        $(
+                            $p => $e,
+                        )+
+                        _ => {
+                            $ctx.errors.push(Error::UnexpectedElement {
+                                xpath: $ctx.xpath.clone(),
+                                name: String::from(name),
+                            });
+                            consume_current_element($ctx);
+                        }
+                    }
+                }
+                XmlEvent::EndElement { .. } => {
+                    $ctx.pop_element();
+                    break;
+                }
+                _ => {}
+            }
+        }
+    };
+}
+
+
+//--------------------------------------------------------------------------------------------------
+pub struct ParseCtx<R: Read> {
+    pub events: XmlEvents<R>,
+    pub xpath: String,
+    pub errors: Vec<Error>,
+}
+
+impl<R: Read> ParseCtx<R> {
+    pub fn push_element(&mut self, name: &str) {
+        self.xpath.push('/');
+        self.xpath.push_str(name);
+    }
+
+    pub fn pop_element(&mut self) {
+        if let Some(separator_pos) = self.xpath.rfind('/') {
+            self.xpath.truncate(separator_pos);
+        } else {
+            self.errors.push(Error::Internal {
+                desc: "ParseCtx push_element/pop_element mismatch.",
+            });
+        }
+    }
+}
+
+pub fn parse_integer<R: Read>(ctx: &mut ParseCtx<R>, text: &str) -> Option<i64> {
+    let parse_res = if text.starts_with("0x") {
+        i64::from_str_radix(text.split_at(2).1, 16)
+    } else {
+        i64::from_str_radix(text, 10)
+    };
+
+    if let Ok(v) = parse_res {
+        Some(v)
+    } else {
+        ctx.errors.push(Error::SchemaViolation {
+            xpath: ctx.xpath.clone(),
+            desc: format!("Value '{}' is not valid base 10 or 16 integer.", text),
+        });
+        None
+    }
+}
+
+pub fn parse_int_attribute<I: FromStr<Err = std::num::ParseIntError>, R: Read>(
+    ctx: &mut ParseCtx<R>,
+    text: String,
+    attribute_name: &str,
+) -> Option<I> {
+    match I::from_str(&text) {
+        Ok(v) => Some(v),
+        Err(e) => {
+            ctx.errors.push(Error::ParseIntError {
+                xpath: xpath_attribute(&ctx.xpath, attribute_name),
+                text: text,
+                error: e,
+            });
+            None
+        }
+    }
+}
+
+pub fn consume_current_element<R: Read>(ctx: &mut ParseCtx<R>) {
+    let mut depth = 1;
+    while let Some(Ok(e)) = ctx.events.next() {
+        match e {
+            XmlEvent::StartElement { name, .. } => {
+                ctx.push_element(name.local_name.as_str());
+                depth += 1;
+            }
+            XmlEvent::EndElement { .. } => {
+                depth -= 1;
+                ctx.pop_element();
+                if depth == 0 {
+                    break;
+                }
+            }
+            _ => (),
+        }
+    }
+}
+
+pub fn parse_text_element<R: Read>(ctx: &mut ParseCtx<R>) -> String {
+    let mut result = String::new();
+    let mut depth = 1;
+    while let Some(Ok(e)) = ctx.events.next() {
+        match e {
+            XmlEvent::StartElement { name, .. } => {
+                ctx.push_element(name.local_name.as_str());
+                depth += 1;
+            }
+            XmlEvent::Characters(text) => result.push_str(&text),
+            XmlEvent::EndElement { .. } => {
+                depth -= 1;
+                ctx.pop_element();
+                if depth == 0 {
+                    break;
+                }
+            }
+            _ => (),
+        }
+    }
+    result
+}
