+++ vk-parse/src/vk/convert.rs
@@ -1,9 +1,11 @@
 extern crate vkxml;
 
 use c;
-use parse::*;
 use std;
+use parse::*;
 use types::*;
+use vk::parse::*;
+use vk::types::*;
 
 //--------------------------------------------------------------------------------------------------
 fn new_field() -> vkxml::Field {
